requirements.txt
pymongo==3.10.1
redis==3.4.1
Flask==1.1.1
gunicorn==20.0.4
python-slugify==4.0.0
Flask-PyMongo==2.3.0
flask-caching==1.8.0